Refactor template to own file and add seed to template eval (#142)
