+++ lib/parser/src/cyber_skyline/chall_parser/__init__.py
@@ -23,7 +23,7 @@ Docker Compose parser with CTF challenge extensions.
 
 from cyber_skyline.chall_parser.yaml_parser import ComposeYamlParser, parse_compose_file, parse_compose_string
 from cyber_skyline.chall_parser.compose import ComposeFile, Service, ChallengeInfo
-from cyber_skyline.chall_parser.rewriter import Template
+from cyber_skyline.chall_parser.template import Template
 
 __all__ = [
     'ComposeYamlParser',
