+++ lib/parser/src/cyber_skyline/chall_parser/yaml_parser.py
@@ -26,7 +26,8 @@ from cattrs.gen import make_dict_structure_fn, make_dict_unstructure_fn
 from cyber_skyline.chall_parser.strategies import configure_union_passthrough, configure_tagged_union
 from typing import get_origin, get_args
 from cyber_skyline.chall_parser.compose import ComposeFile, TextBody
-from cyber_skyline.chall_parser.rewriter import Rewriter, Template
+from cyber_skyline.chall_parser.rewriter import Rewriter
+from cyber_skyline.chall_parser.template import Template
 
 logger = logging.getLogger(__name__)
