+++ tools/chall-check/pyproject.toml
@@ -24,7 +24,7 @@ requires = ["setuptools"]
 build-backend = "setuptools.build_meta"
 
 [tool.uv.sources]
-cyber-skyline-chall-parser = { path = "../../lib/parser" }
+cyber-skyline-chall-parser = { path = "../../lib/parser", editable = true }
 
 [tool.ruff]
 target-version = "py311"
